requirements.txt
acdh-transkribus-utils>=1.1.0
certifi>=2019.3.9
chardet>=3.0.4
Django>=3.0.2
django-autocomplete-light>=3.3.5
django-crispy-forms>=1.7.2
django-filter>=2.1.0
django-leaflet>=0.24.0
django-reversion>=3.0.4
django-super-deduper>=0.1.2
django-tables2>=2.0.6
djangorestframework>=3.9.4
idna>=2.8
isodate>=0.6.0
numpy>=1.16.4
pandas>=0.24.2
pyparsing>=2.4.0
python-dateutil>=2.8.0
pytz>=2019.1
rdflib>=4.2.2
regex>=2018.1.10
requests>=2.22.0
rita>=0.0.7
six>=1.12.0
sqlparse>=0.3.0
ujson>=1.35
urllib3>=1.25.3
wincertstore>=0.2
gunicorn>=20.0.4
